shell-desktop.h: Avoid mass activation of applications from application_id event
I failed to realize that the application_id event is sent at binding,
which is not an issue when launcher is the only application running,
but obviously a big one if multiple surfaces are already present, and
launcher is re-started (so not really something we test).
Added also a note to say a few words about that.

We already activate by default surfaces, and we embedded the activation
call-out into the agl_shell_desktop protocol in that tapShortcut QML, so
we don't actually need further handling. This was just a toy example.

Fix indentation while at it.

Bug-AGL: SPEC-3464

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I46889c5fabdb3936089301b92207b896e4c47ae3